Add `window_attach_buffer` & `window_commit` commands to Compositor API
We're drawing!

+++ libz/shm.c
@@ -1,10 +1,10 @@
 #include "shm.h"
 #include "syscall.h"
 
-uint64_t shm_open(const char *path, uint64_t size, handle_t *fd) {
+int shm_open(const char *path, uint64_t size, handle_t *fd) {
     return syscall_shm_open(path, size, fd);
 }
 
-uint64_t shm_unlink(const char *path) {
+int shm_unlink(const char *path) {
     return syscall_shm_unlink(path);
 }
