+++ libz/shm.h
@@ -3,5 +3,5 @@
 #include <stdint.h>
 #include <libz/handle.h>
 
-uint64_t shm_open(const char *path, uint64_t size, handle_t *fd);
-uint64_t shm_unlink(const char *path);
+int shm_open(const char *path, uint64_t size, handle_t *fd);
+int shm_unlink(const char *path);
